Add helper function for project root directory.
