+++ test/benchmarks/run_benchmarks.dart
@@ -1,14 +1,15 @@
 import 'dart:convert' show JsonEncoder;
-import 'dart:io';
 
 import 'package:web_benchmarks/server.dart';
+
+import 'project_root_directory.dart';
 
 /// Runs the Gallery web benchmarks and reports the benchmark data.
 Future<void> main() async {
   print('Starting web benchmark tests ...');
 
   final taskResult = await serveWebBenchmark(
-    benchmarkAppDirectory: Directory.current.absolute,
+    benchmarkAppDirectory: projectRootDirectory(),
     entryPoint: 'test/benchmarks/client.dart',
     useCanvasKit: false,
   );
